Reorganize Auth0. Create CSEUtils. Create AuthUtils.
