+++ Products/ZenModel/migrate/addAuth0.py
@@ -1,16 +1,16 @@
 ##############################################################################
-# 
+#
 # Copyright (C) Zenoss, Inc. 2018, all rights reserved.
-# 
+#
 # This content is made available according to terms specified in
 # License.zenoss under the directory where your Zenoss product is installed.
-# 
+#
 ##############################################################################
 
 
 __doc__='''
 This migration script adds Auth0 plugin to PAS.
-''' 
+'''
 
 import logging
 from Products.ZenModel.ZMigrateVersion import (
@@ -19,7 +19,7 @@ from Products.ZenModel.ZMigrateVersion import (
 log = logging.getLogger("zen.migrate")
 
 import Migrate
-from Products.ZenUtils.Auth0 import setup
+from Products.ZenUtils.Auth0.Auth0 import setup
 
 
 class AddAuth0(Migrate.Step):
