+++ Products/ZenUtils/Auth0/configure.zcml
@@ -1,10 +1,11 @@
+<?xml version="1.0"?>
 <configure
     xmlns="http://namespaces.zope.org/zope"
     xmlns:browser="http://namespaces.zope.org/browser"
     xmlns:five="http://namespaces.zope.org/five">
 
 <browser:page
-    name="callback"
+    name="Auth0Callback"
     for="*"
     class=".Auth0Callback.Auth0Callback"
     permission="zope2.Public"
