+++ Products/ZenUtils/Utils.py
@@ -1,6 +1,6 @@
 ##############################################################################
 #
-# Copyright (C) Zenoss, Inc. 2007, all rights reserved.
+# Copyright (C) Zenoss, Inc. 2007, 2018 all rights reserved.
 #
 # This content is made available according to terms specified in
 # License.zenoss under the directory where your Zenoss product is installed.
@@ -2217,3 +2217,14 @@ def escapeSpecChars(value):
     escape_re = re.compile(r'(?<!\\)(?P<char>[$&|+\-!(){}[\]^~*?:])')
     return escape_re.sub(r'\\\g<char>', value)
 
+def getQueryArgsFromRequest(request):
+    """Returns a map of query args created from a zope HTTPRequest object
+    """
+    query_args = {}
+    for arg in request.QUERY_STRING.split('&'):
+        parts = arg.split('=', 1)
+        if len(parts) == 2:
+            query_args[parts[0]] = parts[1]
+        elif len(parts) == 1 and parts[0]:
+            query_args[parts[0]] = True
+    return query_args
